Conflicts:

	debian/control

+++ src/rbd.cc
@@ -200,7 +200,7 @@ static int do_rollback_snap(librbd::Image& image, const char *snapname)
 
 static int export_read_cb(uint64_t ofs, size_t len, const char *buf, void *arg)
 {
-  int ret;
+  ssize_t ret;
   int fd = *(int *)arg;
 
   if (!buf) /* a hole */
@@ -219,7 +219,7 @@ static int export_read_cb(uint64_t ofs, size_t len, const char *buf, void *arg)
 
 static int do_export(librbd::Image& image, const char *path)
 {
-  int r;
+  int64_t r;
   librbd::image_info_t info;
   bufferlist bl;
   int fd = open(path, O_WRONLY | O_CREAT | O_EXCL, 0644);
